Fix a signed integer overflow problem in the testcase() macro.
FossilOrigin-Name: f18f5f5893a0d37ecedcfcfdddbe06cdbe0b4243

+++ src/util.c
@@ -26,8 +26,8 @@
 */
 #ifdef SQLITE_COVERAGE_TEST
 void sqlite3Coverage(int x){
-  static int dummy = 0;
-  dummy += x;
+  static unsigned dummy = 0;
+  dummy += (unsigned)x;
 }
 #endif
